Llama/requirements.txt
Flask
langchain_groq
python-dotenv
groq
Flask-Cors
gunicorn
joblib
pandas
scikit-learn
numpy
OpenAI